pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anyrun-tg-bot"
version = "0.7.2"
description = "A Telegram bot for interacting with the ANY.RUN API"
authors = [
    { name = "Kirill Bykov", email = "me@bykovk.pro" }
]
readme = "README.md"
license = { text = "MIT License" }
keywords = ["anyrun", "any.run", "telegram", "bot", "api", "malware", "analysis", "security", "sandbox"]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent"
]
dependencies = [
    "python-telegram-bot>=21.6",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.10.10",
    "validators>=0.34.0",
    "python-dateutil>=2.8.2",
    "packaging>=23.2"
]

[project.urls]
Homepage = "https://any.run"
Repository = "https://github.com/bykovk-pro/anyrun-tg-bot"
